Pushing new repo structure with app package.
