+++ app/abstracts/exceptions_abstract.py
