+++ app/routers/fastapi_router.py
@@ -2,14 +2,14 @@ import os
 from io import BytesIO
 from fastapi import APIRouter
 from fastapi import File, UploadFile
-from abstracts.exceptions_abstract import BaseCustomException
-from schemas.responses import FileUploadResponse, FileDownloadResponse, FileReplaceResponse, FileDeleteResponse, \
+from app.abstracts.exceptions_abstract import BaseCustomException
+from app.schemas.responses import FileUploadResponse, FileDownloadResponse, FileReplaceResponse, FileDeleteResponse, \
     FileRenameResponse, StandardResponse, DatabaseSelectResponse
-from services.aws_database_manager import AWSDatabaseManager
-from services.aws_file_manager import AWSFileManager
-from services.local_database_manager import LocalDatabaseManager
-from services.local_file_manager import LocalFileManager
-from utils.utils import create_file_metadata
+from app.services.aws_database_manager import AWSDatabaseManager
+from app.services.aws_file_manager import AWSFileManager
+from app.services.local_database_manager import LocalDatabaseManager
+from app.services.local_file_manager import LocalFileManager
+from app.utils.utils import create_file_metadata
 
 router = APIRouter()
 FILE_STORAGE_TYPE = os.getenv("FILE_STORAGE_TYPE", "local")
