+++ scripts/launch.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 # launch.sh is a script that launches the application
-uvicorn main:app --reload
+uvicorn app.main:app --reload
